# 3rdparty/openapi.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Sandboxed Agent Platform Gateway",
    "version": "0.3.0"
  },
  "basePath": "/",
  "paths": {
...
    "/api/mcp": {
      "post": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Register one immutable MCP Registry entry version. transport http: url + auth_header/auth_scheme + static headers (contributes one egress whitelist host per Session). transport stdio: command + args + env; the bound credential is injected via credential_env (Sessions selecting any stdio server run without egress restriction).",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "MCP Registry"
        ],
        "summary": "Register an MCP entry version",
        "operationId": "createMCPEntry",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateMCPEntryReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/MCPEntryResp"
            }
          }
        }
      }
    },
    "/api/mcp/list": {
      "get": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Default: unarchived public view across all owners. scope=mine: only the caller's own entries, including archived ones.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "MCP Registry"
        ],
        "summary": "List MCP entries",
        "operationId": "listMCPEntries",
        "parameters": [
          {
            "type": "string",
            "name": "scope",
            "in": "query",
            "allowEmptyValue": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListMCPEntriesResp"
            }
          }
        }
      }
    },
...
    "/api/my/agents": {
      "get": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "The caller's own managed Agents with full configuration (suitable for editing).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "My Agents"
        ],
        "summary": "List my agents",
        "operationId": "listMyAgents",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListMyAgentsResp"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Self-service creation of a v2 managed Agent owned by the caller. Accepts the same managed fields as /agent/register (engine, instructions, start_prompt_template, skills, mcp_bindings, credential_slots, subagents, ...).",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "My Agents"
        ],
        "summary": "Create a managed agent",
        "operationId": "createMyAgent",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterAgentReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/RegisterAgentResp"
            }
          }
        }
      }
    },
    "/api/my/agents/{agentId}": {
      "put": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Merge-update one of the caller's managed Agents: omitted fields keep stored values; non-nil arrays/objects replace the whole set. Effective configuration changes mint a new immutable Agent Version.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "My Agents"
        ],
        "summary": "Update a managed agent",
        "operationId": "updateMyAgent",
        "parameters": [
          {
            "type": "string",
            "name": "agentId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateMyAgentReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/RegisterAgentResp"
            }
          }
        }
      }
    },
...
    "/api/skill/list": {
      "get": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Default: unarchived public view across all owners. scope=mine: only the caller's own entries, including archived ones.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "Skills"
        ],
        "summary": "List skills",
        "operationId": "listSkills",
        "parameters": [
          {
            "type": "string",
            "name": "scope",
            "in": "query",
            "allowEmptyValue": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListSkillsResp"
            }
          }
        }
      }
    },
...
    "/api/task/submit": {
      "post": {
        "security": [
          {
            "apiKey": []
          }
        ],
        "description": "Queue a run of an Agent. For v1 image agents this creates a Task; if agent_id points to a v2 managed Agent this transparently creates a Session instead (task_id == session_id, params map to start prompt values, input_files become Session File Resources).",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Submit a task",
        "operationId": "submitTask",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitTaskReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/SubmitTaskResp"
            }
          }
        }
      }
    },
...
    "/api/task/{taskId}/result": {
      "get": {
        "description": "Same shape as /status, oriented to the result page (includes archive info). Owner or public; auth optional.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Get task result",
        "operationId": "getTaskResult",
        "parameters": [
          {
            "type": "string",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/TaskStatusResp"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateMCPEntryReq": {
      "type": "object",
      "required": [
        "slug",
        "version"
      ],
      "properties": {
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "auth_header": {
          "type": "string"
        },
        "auth_scheme": {
          "type": "string"
        },
        "command": {
          "type": "string"
        },
        "credential_env": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "headers": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "requires_credential": {
          "type": "boolean"
        },
        "slug": {
          "type": "string"
        },
        "transport": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
...
    "CredentialSlotDef": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "required": {
          "type": "boolean"
        }
      }
    },
...
    "ListMCPEntriesResp": {
      "type": "object",
      "required": [
        "entries"
      ],
      "properties": {
        "entries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MCPEntryResp"
          }
        }
      }
    },
...
    "ListMyAgentsResp": {
      "type": "object",
      "required": [
        "agents"
      ],
      "properties": {
        "agents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MyAgentResp"
          }
        }
      }
    },
...
    "ListSkillsResp": {
      "type": "object",
      "required": [
        "skills"
      ],
      "properties": {
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SkillResp"
          }
        }
      }
    },
...
    "MCPEntryResp": {
      "type": "object",
      "required": [
        "entry_id",
        "slug",
        "version",
        "name",
        "transport",
        "requires_credential",
        "archived",
        "created_at"
      ],
      "properties": {
        "archived": {
          "type": "boolean"
        },
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "auth_header": {
          "type": "string"
        },
        "auth_scheme": {
          "type": "string"
        },
        "command": {
          "type": "string"
        },
        "created_at": {
          "type": "integer"
        },
        "credential_env": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "entry_id": {
          "type": "string"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "headers": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "requires_credential": {
          "type": "boolean"
        },
        "slug": {
          "type": "string"
        },
        "transport": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
...
    "MyAgentResp": {
      "type": "object",
      "required": [
        "agent_id",
        "name",
        "is_public",
        "timeout",
        "engine",
        "current_version_id",
        "archived"
      ],
      "properties": {
        "agent_id": {
          "type": "string"
        },
        "archived": {
          "type": "boolean"
        },
        "credential_slots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CredentialSlotDef"
          }
        },
        "current_version_id": {
          "type": "integer"
        },
        "default_bindings": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "default_model_id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "engine": {
          "type": "string"
        },
        "harness_image": {
          "type": "string"
        },
        "instructions": {
          "type": "string"
        },
        "is_public": {
          "type": "boolean"
        },
        "mcp_bindings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MCPBindingDef"
          }
        },
        "name": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SkillRef"
          }
        },
        "start_prompt_template": {
          "type": "string"
        },
        "subagents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubagentRef"
          }
        },
        "timeout": {
          "type": "integer"
        }
      }
    },
...
    "RegisterAgentReq": {
      "type": "object",
      "properties": {
        "accepted_file_types": {
          "type": "string"
        },
        "agent_id": {
          "type": "string"
        },
        "credential_slots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CredentialSlotDef"
          }
        },
        "default_bindings": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "default_model_id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "engine": {
          "type": "string"
        },
        "entrypoint": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "env": {
          "type": "object",
          "additionalProperties": {}
        },
        "harness_image": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "input_schema": {
          "type": "object",
          "additionalProperties": {}
        },
        "instructions": {
          "type": "string"
        },
        "mcp_bindings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MCPBindingDef"
          }
        },
        "mcp_servers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MCPServerDef"
          }
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "metrics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AgentMetricDef"
          }
        },
        "name": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SkillRef"
          }
        },
        "start_prompt_template": {
          "type": "string"
        },
        "subagents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubagentRef"
          }
        },
        "timeout": {
          "type": "integer"
        }
      }
    },
    "RegisterAgentResp": {
      "type": "object",
      "required": [
        "agent_id"
      ],
      "properties": {
        "agent_id": {
          "type": "string"
        },
        "managed_version": {
          "type": "integer"
        }
      }
    },
...
    "SubmitTaskReq": {
      "type": "object",
      "required": [
        "agent_id"
      ],
      "properties": {
        "agent_id": {
          "type": "string"
        },
        "input_files": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "model_id": {
          "type": "string"
        },
        "params": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "SubmitTaskResp": {
      "type": "object",
      "required": [
        "task_id",
        "status"
      ],
      "properties": {
        "model_id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "task_id": {
          "type": "string"
        }
      }
    },
...
    "UpdateMyAgentReq": {
      "type": "object",
      "properties": {
        "credential_slots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CredentialSlotDef"
          }
        },
        "default_bindings": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "default_model_id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "engine": {
          "type": "string"
        },
        "env": {
          "type": "object",
          "additionalProperties": {}
        },
        "harness_image": {
          "type": "string"
        },
        "input_schema": {
          "type": "object",
          "additionalProperties": {}
        },
        "instructions": {
          "type": "string"
        },
        "mcp_bindings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MCPBindingDef"
          }
        },
        "name": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SkillRef"
          }
        },
        "start_prompt_template": {
          "type": "string"
        },
        "subagents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubagentRef"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "description": "Bearer token from POST /api/auth/login. Format: 'Bearer \u003ctoken\u003e'. SSE endpoints alternatively accept '?token=\u003ctoken\u003e' as a query parameter.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}